// frontend/src/api/generated/models/receiptRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable ticket tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */

export interface ReceiptRead {
  filename: string;
  content_type: string;
  size_bytes: number;
  is_image: boolean;
  url: string;
}